+++ feersum_nlu/api/__init__.py
@@ -3,11 +3,13 @@ from __future__ import absolute_import
 # flake8: noqa
 
 # import apis into api package
+from feersum_nlu.api.crf_entity_extractors_api import CrfEntityExtractorsApi
 from feersum_nlu.api.dashboard_api import DashboardApi
 from feersum_nlu.api.data_objects_api import DataObjectsApi
 from feersum_nlu.api.date_parsers_api import DateParsersApi
 from feersum_nlu.api.duckling_entity_extractors_api import DucklingEntityExtractorsApi
 from feersum_nlu.api.faq_matchers_api import FaqMatchersApi
+from feersum_nlu.api.health_api import HealthApi
 from feersum_nlu.api.intent_classifiers_api import IntentClassifiersApi
 from feersum_nlu.api.language_recognisers_api import LanguageRecognisersApi
 from feersum_nlu.api.person_name_entity_extractors_api import PersonNameEntityExtractorsApi