+++ feersum_nlu/models/duckling_entity_extractor_instance_detail.py
@@ -35,6 +35,7 @@ class DucklingEntityExtractorInstanceDetail(object):
         'id': 'str',
         'long_name': 'str',
         'desc': 'str',
+        'readonly': 'bool',
         'reference_time': 'str'
     }
 
@@ -43,16 +44,18 @@ class DucklingEntityExtractorInstanceDetail(object):
         'id': 'id',
         'long_name': 'long_name',
         'desc': 'desc',
+        'readonly': 'readonly',
         'reference_time': 'reference_time'
     }
 
-    def __init__(self, name=None, id=None, long_name=None, desc=None, reference_time=None):  # noqa: E501
+    def __init__(self, name=None, id=None, long_name=None, desc=None, readonly=None, reference_time=None):  # noqa: E501
         """DucklingEntityExtractorInstanceDetail - a model defined in Swagger"""  # noqa: E501
 
         self._name = None
         self._id = None
         self._long_name = None
         self._desc = None
+        self._readonly = None
         self._reference_time = None
         self.discriminator = None
 
@@ -62,6 +65,8 @@ class DucklingEntityExtractorInstanceDetail(object):
             self.long_name = long_name
         if desc is not None:
             self.desc = desc
+        if readonly is not None:
+            self.readonly = readonly
         if reference_time is not None:
             self.reference_time = reference_time
 
@@ -160,6 +165,29 @@ class DucklingEntityExtractorInstanceDetail(object):
         """
 
         self._desc = desc
+
+    @property
+    def readonly(self):
+        """Gets the readonly of this DucklingEntityExtractorInstanceDetail.  # noqa: E501
+
+        Indicates if the model is readonly and not editable.  # noqa: E501
+
+        :return: The readonly of this DucklingEntityExtractorInstanceDetail.  # noqa: E501
+        :rtype: bool
+        """
+        return self._readonly
+
+    @readonly.setter
+    def readonly(self, readonly):
+        """Sets the readonly of this DucklingEntityExtractorInstanceDetail.
+
+        Indicates if the model is readonly and not editable.  # noqa: E501
+
+        :param readonly: The readonly of this DucklingEntityExtractorInstanceDetail.  # noqa: E501
+        :type: bool
+        """
+
+        self._readonly = readonly
 
     @property
     def reference_time(self):